requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
puremagic==1.5
opsdroid-get-image-size==0.2.2
gmqtt==0.3.7
motor==2.0.0
docker==4.0.1
minio==4.0.18
rasa==1.0.3
rasa_sdk==1.0.0
openpyxl==2.6.2
nltk==3.4.3
pattern==3.6
PyYaml